update generator
cast id to int

+++ RRScheduler/src/rr/Generator.java
@@ -36,8 +36,8 @@ public class Generator {
 	public void run(){
 		//add process to end of queue
 		double execTime = (Math.random()*12) + 12; 
-		double id = (Math.random()*12) + 12;
-		Process pro = new Process(execTime, id); //pass in arguements of time 
+		int id = (int)((Math.random()*12) + 12);
+		Process pro = new Process(execTime,  id); //pass in arguements of time 
 		ReadyQueue.addProcess(pro);
 	}
